made optimizations to the sensor chart tab and adapted to changes made in open_earable_flutter.
